refactor(test-runner-commands): migrate tests from mocha to node:test
Replace mocha globals with node:test across 10 test files. No chai was
used -- tests rely on runTests throwing on failure.

- Add `import { describe, it } from 'node:test'` (+ before/after where used)
- `../../src/*.js` -> `../../dist/*.js`
- `../selenium-server.js` -> `../selenium-server.ts`
- `__dirname` -> `import.meta.dirname`
- `this.timeout(N)` -> `{ timeout: N }` on describe
- Add `--experimental-strip-types` for CI compat
