+++ packages/test-runner-commands/test/a11y-snapshot/a11ySnapshotPlugin.test.ts
@@ -1,16 +1,15 @@
+import { describe, it } from 'node:test';
 import path from 'path';
 import { runTests } from '@web/test-runner-core/test-helpers';
 import { chromeLauncher } from '@web/test-runner-chrome';
 import { playwrightLauncher } from '@web/test-runner-playwright';
 
-import { a11ySnapshotPlugin } from '../../src/a11ySnapshotPlugin.js';
+import { a11ySnapshotPlugin } from '../../dist/a11ySnapshotPlugin.js';
 
-describe('a11ySnapshotPlugin', function test() {
-  this.timeout(20000);
-
+describe('a11ySnapshotPlugin', { timeout: 20000 }, () => {
   it('can find accessibility nodes in the returned accessibility tree on puppeteer', async () => {
     await runTests({
-      files: [path.join(__dirname, 'browser-test.js')],
+      files: [path.join(import.meta.dirname, 'browser-test.js')],
       browsers: [chromeLauncher()],
       plugins: [a11ySnapshotPlugin()],
     });
@@ -18,7 +17,7 @@ describe('a11ySnapshotPlugin', function test() {
 
   it('can find accessibility nodes in the returned accessibility tree on playwright', async () => {
     await runTests({
-      files: [path.join(__dirname, 'browser-test.js')],
+      files: [path.join(import.meta.dirname, 'browser-test.js')],
       browsers: [
         playwrightLauncher({ product: 'chromium' }),
         playwrightLauncher({ product: 'firefox' }),
