+++ packages/test-runner-commands/test/emulate-media/emulateMediaPlugin.test.ts
@@ -1,18 +1,17 @@
+import { describe, it } from 'node:test';
 import path from 'path';
 import { runTests } from '@web/test-runner-core/test-helpers';
 import { chromeLauncher } from '@web/test-runner-chrome';
 import { playwrightLauncher } from '@web/test-runner-playwright';
 
-import { emulateMediaPlugin } from '../../src/emulateMediaPlugin.js';
+import { emulateMediaPlugin } from '../../dist/emulateMediaPlugin.js';
 
-describe('emulateMediaPlugin', function test() {
-  this.timeout(20000);
-
+describe('emulateMediaPlugin', { timeout: 20000 }, () => {
   it('can emulate media on puppeteer', async () => {
     await runTests({
       files: [
-        path.join(__dirname, 'browser-test.js'),
-        path.join(__dirname, 'prefers-reduced-motion-test.js'),
+        path.join(import.meta.dirname, 'browser-test.js'),
+        path.join(import.meta.dirname, 'prefers-reduced-motion-test.js'),
       ],
 
       browsers: [chromeLauncher()],
@@ -23,8 +22,8 @@ describe('emulateMediaPlugin', function test() {
   it('can emulate media on playwright', async () => {
     await runTests({
       files: [
-        path.join(__dirname, 'browser-test.js'),
-        path.join(__dirname, 'prefers-reduced-motion-test.js'),
+        path.join(import.meta.dirname, 'browser-test.js'),
+        path.join(import.meta.dirname, 'prefers-reduced-motion-test.js'),
       ],
       browsers: [
         playwrightLauncher({ product: 'chromium' }),
@@ -37,7 +36,7 @@ describe('emulateMediaPlugin', function test() {
 
   it('can emulate forced-colors on playwright, except webkit', async () => {
     await runTests({
-      files: [path.join(__dirname, 'forced-colors-test.js')],
+      files: [path.join(import.meta.dirname, 'forced-colors-test.js')],
       browsers: [
         playwrightLauncher({ product: 'chromium' }),
         playwrightLauncher({ product: 'firefox' }),
