+++ packages/test-runner-commands/test/file/filePlugin.test.ts
@@ -1,15 +1,14 @@
+import { describe, it } from 'node:test';
 import path from 'path';
 import { runTests } from '@web/test-runner-core/test-helpers';
 import { chromeLauncher } from '@web/test-runner-chrome';
 
-import { filePlugin } from '../../src/filePlugin.js';
+import { filePlugin } from '../../dist/filePlugin.js';
 
-describe('filePlugin', function test() {
-  this.timeout(20000);
-
+describe('filePlugin', { timeout: 20000 }, () => {
   it('passes file plugin tests', async () => {
     await runTests({
-      files: [path.join(__dirname, 'browser-test.js')],
+      files: [path.join(import.meta.dirname, 'browser-test.js')],
       browsers: [chromeLauncher()],
       plugins: [filePlugin()],
       logger: {
