+++ packages/test-runner-commands/test/send-keys/sendKeysPlugin.test.ts
@@ -1,16 +1,15 @@
+import { describe, it } from 'node:test';
 import path from 'path';
 import { runTests } from '@web/test-runner-core/test-helpers';
 import { chromeLauncher } from '@web/test-runner-chrome';
 import { playwrightLauncher } from '@web/test-runner-playwright';
 
-import { sendKeysPlugin } from '../../src/sendKeysPlugin.js';
+import { sendKeysPlugin } from '../../dist/sendKeysPlugin.js';
 
-describe('sendKeysPlugin', function test() {
-  this.timeout(20000);
-
+describe('sendKeysPlugin', { timeout: 20000 }, () => {
   it('can send keys on puppeteer', async () => {
     await runTests({
-      files: [path.join(__dirname, 'browser-test.js')],
+      files: [path.join(import.meta.dirname, 'browser-test.js')],
       browsers: [chromeLauncher()],
       plugins: [sendKeysPlugin()],
     });
@@ -18,7 +17,7 @@ describe('sendKeysPlugin', function test() {
 
   it('can send keys on playwright', async () => {
     await runTests({
-      files: [path.join(__dirname, 'browser-test.js')],
+      files: [path.join(import.meta.dirname, 'browser-test.js')],
       browsers: [
         playwrightLauncher({ product: 'chromium' }),
         playwrightLauncher({ product: 'firefox' }),
