+++ packages/test-runner-commands/test/send-mouse/sendMousePlugin.test.ts
@@ -1,18 +1,17 @@
+import { describe, it, before, after } from 'node:test';
 import path from 'path';
 import selenium from 'selenium-standalone';
 import { runTests } from '@web/test-runner-core/test-helpers';
 import { chromeLauncher } from '@web/test-runner-chrome';
 import { webdriverLauncher } from '@web/test-runner-webdriver';
 import { playwrightLauncher } from '@web/test-runner-playwright';
-import { sendMousePlugin } from '../../src/sendMousePlugin.js';
-import { startSeleniumServer } from '../selenium-server.js';
+import { sendMousePlugin } from '../../dist/sendMousePlugin.js';
+import { startSeleniumServer } from '../selenium-server.ts';
 
-describe('sendMousePlugin', function test() {
-  this.timeout(50000);
-
+describe('sendMousePlugin', { timeout: 50000 }, () => {
   it('can send mouse on puppeteer', async () => {
     await runTests({
-      files: [path.join(__dirname, 'browser-test.js')],
+      files: [path.join(import.meta.dirname, 'browser-test.js')],
       browsers: [chromeLauncher()],
       plugins: [sendMousePlugin()],
     });
@@ -20,7 +19,7 @@ describe('sendMousePlugin', function test() {
 
   it('can send mouse on playwright', async () => {
     await runTests({
-      files: [path.join(__dirname, 'browser-test.js')],
+      files: [path.join(import.meta.dirname, 'browser-test.js')],
       browsers: [
         playwrightLauncher({ product: 'chromium' }),
         playwrightLauncher({ product: 'firefox' }),
@@ -52,7 +51,7 @@ describe('sendMousePlugin', function test() {
 
     it('can send mouse on webdriver', async () => {
       await runTests({
-        files: [path.join(__dirname, 'browser-test.js')],
+        files: [path.join(import.meta.dirname, 'browser-test.js')],
         concurrency: 1,
         browsers: [
           webdriverLauncher({
