+++ packages/test-runner-commands/test/set-viewport/setViewportPlugin.test.ts
@@ -1,17 +1,16 @@
+import { describe, it } from 'node:test';
 import path from 'path';
 
 import { runTests } from '@web/test-runner-core/test-helpers';
 import { chromeLauncher } from '@web/test-runner-chrome';
 import { playwrightLauncher } from '@web/test-runner-playwright';
 
-import { setViewportPlugin } from '../../src/setViewportPlugin.js';
+import { setViewportPlugin } from '../../dist/setViewportPlugin.js';
 
-describe('setViewportPlugin', function test() {
-  this.timeout(20000);
-
+describe('setViewportPlugin', { timeout: 20000 }, () => {
   it('can set the viewport on puppeteer', async () => {
     await runTests({
-      files: [path.join(__dirname, 'browser-test.js')],
+      files: [path.join(import.meta.dirname, 'browser-test.js')],
       browsers: [chromeLauncher()],
       plugins: [setViewportPlugin()],
     });
@@ -19,7 +18,7 @@ describe('setViewportPlugin', function test() {
 
   it('can set the viewport on playwright', async () => {
     await runTests({
-      files: [path.join(__dirname, 'browser-test.js')],
+      files: [path.join(import.meta.dirname, 'browser-test.js')],
       browsers: [
         playwrightLauncher({ product: 'chromium' }),
         playwrightLauncher({ product: 'firefox' }),
