+++ packages/test-runner-commands/test/snapshot/snapshotPlugin.test.ts
@@ -1,15 +1,14 @@
+import { describe, it } from 'node:test';
 import path from 'path';
 import { runTests } from '@web/test-runner-core/test-helpers';
 import { playwrightLauncher } from '@web/test-runner-playwright';
 
-import { snapshotPlugin } from '../../src/snapshotPlugin.js';
+import { snapshotPlugin } from '../../dist/snapshotPlugin.js';
 
-describe('snapshotPlugin', function test() {
-  this.timeout(20000);
-
+describe('snapshotPlugin', { timeout: 20000 }, () => {
   it('passes snapshot tests', async () => {
     await runTests({
-      files: [path.join(__dirname, 'browser-test.js')],
+      files: [path.join(import.meta.dirname, 'browser-test.js')],
       browsers: [
         playwrightLauncher({ product: 'firefox' }),
         playwrightLauncher({ product: 'chromium' }),
@@ -19,7 +18,7 @@ describe('snapshotPlugin', function test() {
     });
 
     await runTests({
-      files: [path.join(__dirname, 'src', 'nested-test.js')],
+      files: [path.join(import.meta.dirname, 'src', 'nested-test.js')],
       browsers: [
         playwrightLauncher({ product: 'firefox' }),
         playwrightLauncher({ product: 'chromium' }),
